Remove dependency on polymer as it breaks cocoonjs
Also fix canvas size to integer (also breaks cocoonjs).

+++ web/index.dart
@@ -8,8 +8,8 @@ void main() {
 
   canvas
       ..setAttribute("screencanvas", "true")
-      ..width = window.innerWidth
-      ..height = window.innerHeight;
+      ..width = window.innerWidth.toInt()
+      ..height = window.innerHeight.toInt();
 
   document.body.children.add(canvas);
